GraphQLClient now runs spring boot server for junit tests
Updated maven plugin for spring-boot to not overwrite the jar file, but
create a new bootable jar file

Updated test to run as a spring boot test using a random port for the
server config
