+++ crm-graphql-spring-boot-jwt-server/pom.xml
@@ -61,6 +61,12 @@
 
 		<!-- Test Dependencies -->
 		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-test</artifactId>
+			<scope>test</scope>
+		</dependency>
+
+		<dependency>
 			<groupId>org.apache.httpcomponents</groupId>
 			<artifactId>httpclient</artifactId>
 			<scope>test</scope>
@@ -72,6 +78,9 @@
 			<plugin>
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
+				<configuration>
+					<classifier>bootable</classifier>
+				</configuration>
 			</plugin>
 		</plugins>
 	</build>
